feat(install): run config validation at the beginning of the install command
