+++ internal/diagnose/mock_config_validator.go
@@ -1,0 +1,13 @@
+package diagnose
+
+import "context"
+
+type MockConfigValidator struct{}
+
+func NewMockConfigValidator() *MockConfigValidator {
+	return &MockConfigValidator{}
+}
+
+func (v *MockConfigValidator) ValidateConfig(ctx context.Context) error {
+	return nil
+}
