+++ internal/install/recipe_installer_test.go
@@ -13,6 +13,7 @@ import (
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
 
+	"github.com/newrelic/newrelic-cli/internal/diagnose"
 	"github.com/newrelic/newrelic-cli/internal/install/discovery"
 	"github.com/newrelic/newrelic-cli/internal/install/execution"
 	"github.com/newrelic/newrelic-cli/internal/install/recipes"
@@ -40,6 +41,7 @@ var (
 	p               = ux.NewMockPrompter()
 	pi              = ux.NewMockProgressIndicator()
 	lkf             = NewMockLicenseKeyFetcher()
+	cv              = diagnose.NewMockConfigValidator()
 )
 
 func TestInstall(t *testing.T) {
@@ -56,7 +58,7 @@ func TestNewRecipeInstaller_InstallerContextFields(t *testing.T) {
 		SkipApm:            true,
 	}
 
-	i := RecipeInstaller{ic, d, l, mv, f, e, v, ff, status, p, pi, lkf}
+	i := RecipeInstaller{ic, d, l, mv, f, e, v, ff, status, p, pi, lkf, cv}
 
 	require.True(t, reflect.DeepEqual(ic, i.InstallerContext))
 }
@@ -65,7 +67,7 @@ func TestShouldGetRecipeFromURL(t *testing.T) {
 	ic := InstallerContext{}
 	ff = recipes.NewMockRecipeFileFetcher()
 	ff.FetchRecipeFileFunc = fetchRecipeFileFunc
-	i := RecipeInstaller{ic, d, l, mv, f, e, v, ff, status, p, pi, lkf}
+	i := RecipeInstaller{ic, d, l, mv, f, e, v, ff, status, p, pi, lkf, cv}
 
 	recipe, err := i.recipeFromPath("http://recipe/URL")
 	require.NoError(t, err)
@@ -77,7 +79,7 @@ func TestShouldGetRecipeFromFile(t *testing.T) {
 	ic := InstallerContext{}
 	ff = recipes.NewMockRecipeFileFetcher()
 	ff.LoadRecipeFileFunc = loadRecipeFileFunc
-	i := RecipeInstaller{ic, d, l, mv, f, e, v, ff, status, p, pi, lkf}
+	i := RecipeInstaller{ic, d, l, mv, f, e, v, ff, status, p, pi, lkf, cv}
 
 	recipe, err := i.recipeFromPath("file.txt")
 	require.NoError(t, err)
@@ -92,7 +94,7 @@ func TestInstall_Basic(t *testing.T) {
 		{Name: types.InfraAgentRecipeName},
 		{Name: types.LoggingRecipeName},
 	}
-	i := RecipeInstaller{ic, d, l, mv, f, e, v, ff, status, p, pi, lkf}
+	i := RecipeInstaller{ic, d, l, mv, f, e, v, ff, status, p, pi, lkf, cv}
 	err := i.Install()
 	require.NoError(t, err)
 	require.Equal(t, f.FetchRecipeNameCount[types.InfraAgentRecipeName], 1)
@@ -112,7 +114,7 @@ func TestInstall_DiscoveryComplete(t *testing.T) {
 		},
 	}
 
-	i := RecipeInstaller{ic, d, l, mv, f, e, v, ff, status, p, pi, lkf}
+	i := RecipeInstaller{ic, d, l, mv, f, e, v, ff, status, p, pi, lkf, cv}
 
 	err := i.Install()
 	require.NoError(t, err)
@@ -135,7 +137,7 @@ func TestInstall_FailsOnInvalidOs(t *testing.T) {
 		},
 	}
 
-	i := RecipeInstaller{ic, discover, l, mv, f, e, v, ff, status, p, pi, lkf}
+	i := RecipeInstaller{ic, discover, l, mv, f, e, v, ff, status, p, pi, lkf, cv}
 
 	err := i.Install()
 	require.Error(t, err)
@@ -163,7 +165,7 @@ func TestInstall_RecipesAvailable(t *testing.T) {
 		},
 	}
 
-	i := RecipeInstaller{ic, d, l, mv, f, e, v, ff, status, p, pi, lkf}
+	i := RecipeInstaller{ic, d, l, mv, f, e, v, ff, status, p, pi, lkf, cv}
 	err := i.Install()
 	require.NoError(t, err)
 	require.Equal(t, 1, statusReporters[0].(*execution.MockStatusReporter).RecipesAvailableCallCount)
@@ -205,7 +207,7 @@ func TestInstall_RecipeInstalled(t *testing.T) {
 
 	v = validation.NewMockRecipeValidator()
 
-	i := RecipeInstaller{ic, d, l, mv, f2, e, v, ff, status, p, pi, lkf}
+	i := RecipeInstaller{ic, d, l, mv, f2, e, v, ff, status, p, pi, lkf, cv}
 	err := i.Install()
 	require.NoError(t, err)
 	require.Equal(t, 3, statusReporters[0].(*execution.MockStatusReporter).RecipeInstalledCallCount)
@@ -244,7 +246,7 @@ func TestInstall_RecipeFailed(t *testing.T) {
 	v = validation.NewMockRecipeValidator()
 	v.ValidateErr = errors.New("validationErr")
 
-	i := RecipeInstaller{ic, d, l, mv, f, e, v, ff, status, p, pi, lkf}
+	i := RecipeInstaller{ic, d, l, mv, f, e, v, ff, status, p, pi, lkf, cv}
 	err := i.Install()
 	require.Error(t, err)
 	require.Equal(t, 1, v.ValidateCallCount)
@@ -269,7 +271,7 @@ func TestInstall_InstallComplete(t *testing.T) {
 
 	v = validation.NewMockRecipeValidator()
 
-	i := RecipeInstaller{ic, d, l, mv, f, e, v, ff, status, p, pi, lkf}
+	i := RecipeInstaller{ic, d, l, mv, f, e, v, ff, status, p, pi, lkf, cv}
 	err := i.Install()
 	require.NoError(t, err)
 	require.Equal(t, 1, statusReporters[0].(*execution.MockStatusReporter).InstallCompleteCallCount)
@@ -288,7 +290,7 @@ func TestInstall_InstallCanceled(t *testing.T) {
 
 	v = validation.NewMockRecipeValidator()
 
-	i := RecipeInstaller{ic, d, l, mv, f, e, v, ff, status, p, pi, lkf}
+	i := RecipeInstaller{ic, d, l, mv, f, e, v, ff, status, p, pi, lkf, cv}
 	err := i.Install()
 	require.Error(t, err)
 	require.Equal(t, 0, statusReporters[0].(*execution.MockStatusReporter).InstallCompleteCallCount)
@@ -317,7 +319,7 @@ func TestInstall_InstallCompleteError(t *testing.T) {
 	v = validation.NewMockRecipeValidator()
 	v.ValidateErr = errors.New("test error")
 
-	i := RecipeInstaller{ic, d, l, mv, f, e, v, ff, status, p, pi, lkf}
+	i := RecipeInstaller{ic, d, l, mv, f, e, v, ff, status, p, pi, lkf, cv}
 	err := i.Install()
 	require.Error(t, err)
 	require.Equal(t, 1, statusReporters[0].(*execution.MockStatusReporter).InstallCompleteCallCount)
@@ -358,7 +360,7 @@ func TestInstall_InstallCompleteError_guidedRecipeFail(t *testing.T) {
 		errors.New("testing error"),
 	}
 
-	i := RecipeInstaller{ic, d, l, mv, f, e, v, ff, status, p, pi, lkf}
+	i := RecipeInstaller{ic, d, l, mv, f, e, v, ff, status, p, pi, lkf, cv}
 	err := i.Install()
 	require.NoError(t, err)
 	require.Equal(t, 1, statusReporters[0].(*execution.MockStatusReporter).InstallCompleteCallCount)
@@ -394,7 +396,7 @@ func TestInstall_RecipeSkipped(t *testing.T) {
 		PromptMultiSelectAll: true,
 	}
 
-	i := RecipeInstaller{ic, d, l, mv, f, e, v, ff, status, p, pi, lkf}
+	i := RecipeInstaller{ic, d, l, mv, f, e, v, ff, status, p, pi, lkf, cv}
 	err := i.Install()
 	require.NoError(t, err)
 	require.Equal(t, 1, statusReporters[0].(*execution.MockStatusReporter).RecipeSkippedCallCount)
@@ -432,7 +434,7 @@ func TestInstall_RecipeSkippedApm(t *testing.T) {
 		PromptMultiSelectAll: true,
 	}
 
-	i := RecipeInstaller{ic, d, l, mv, f, e, v, ff, status, p, pi, lkf}
+	i := RecipeInstaller{ic, d, l, mv, f, e, v, ff, status, p, pi, lkf, cv}
 	err := i.Install()
 	require.NoError(t, err)
 	require.Equal(t, 1, statusReporters[0].(*execution.MockStatusReporter).RecipeSkippedCallCount)
@@ -470,7 +472,7 @@ func TestInstall_RecipeSkippedApmAnyKeyword(t *testing.T) {
 		PromptMultiSelectAll: true,
 	}
 
-	i := RecipeInstaller{ic, d, l, mv, f, e, v, ff, status, p, pi, lkf}
+	i := RecipeInstaller{ic, d, l, mv, f, e, v, ff, status, p, pi, lkf, cv}
 	err := i.Install()
 	require.NoError(t, err)
 	require.Equal(t, 1, statusReporters[0].(*execution.MockStatusReporter).RecipeSkippedCallCount)
@@ -507,7 +509,7 @@ func TestInstall_RecipeSkipped_SkipAll(t *testing.T) {
 		PromptMultiSelectVal: []string{},
 	}
 
-	i := RecipeInstaller{ic, d, l, mv, f, e, v, ff, status, p, pi, lkf}
+	i := RecipeInstaller{ic, d, l, mv, f, e, v, ff, status, p, pi, lkf, cv}
 	err := i.Install()
 	require.NoError(t, err)
 	require.Equal(t, 2, statusReporters[0].(*execution.MockStatusReporter).RecipeSkippedCallCount)
@@ -543,7 +545,7 @@ func TestInstall_RecipeSkipped_MultiSelect(t *testing.T) {
 		PromptMultiSelectVal: []string{testRecipeName},
 	}
 
-	i := RecipeInstaller{ic, d, l, mv, f, e, v, ff, status, p, pi, lkf}
+	i := RecipeInstaller{ic, d, l, mv, f, e, v, ff, status, p, pi, lkf, cv}
 	err := i.Install()
 	require.NoError(t, err)
 	require.Equal(t, 1, statusReporters[0].(*execution.MockStatusReporter).RecipeSkippedCallCount)
@@ -606,7 +608,7 @@ func TestInstall_RecipeRecommended(t *testing.T) {
 		PromptMultiSelectVal: []string{testRecipeName},
 	}
 
-	i := RecipeInstaller{ic, d, l, mv, f, e, v, ff, status, p, pi, lkf}
+	i := RecipeInstaller{ic, d, l, mv, f, e, v, ff, status, p, pi, lkf, cv}
 	err := i.Install()
 	require.NoError(t, err)
 	require.Equal(t, 2, statusReporters[0].(*execution.MockStatusReporter).RecipeSkippedCallCount)
@@ -648,7 +650,7 @@ func TestInstall_RecipeSkipped_AssumeYes(t *testing.T) {
 		PromptYesNoVal: true,
 	}
 
-	i := RecipeInstaller{ic, d, l, mv, f, e, v, ff, status, p, pi, lkf}
+	i := RecipeInstaller{ic, d, l, mv, f, e, v, ff, status, p, pi, lkf, cv}
 	err := i.Install()
 	require.NoError(t, err)
 	require.Equal(t, 0, statusReporters[0].(*execution.MockStatusReporter).RecipeSkippedCallCount)
@@ -674,7 +676,7 @@ func TestInstall_TargetedInstall_InstallsInfraAgent(t *testing.T) {
 
 	v = validation.NewMockRecipeValidator()
 
-	i := RecipeInstaller{ic, d, l, mv, f, e, v, ff, status, p, pi, lkf}
+	i := RecipeInstaller{ic, d, l, mv, f, e, v, ff, status, p, pi, lkf, cv}
 	err := i.Install()
 	require.NoError(t, err)
 	require.Equal(t, 1, statusReporters[0].(*execution.MockStatusReporter).RecipeInstalledCallCount)
@@ -704,7 +706,7 @@ func TestInstall_TargetedInstall_InstallsInfraAgentDependency(t *testing.T) {
 
 	v = validation.NewMockRecipeValidator()
 
-	i := RecipeInstaller{ic, d, l, mv, f, e, v, ff, status, p, pi, lkf}
+	i := RecipeInstaller{ic, d, l, mv, f, e, v, ff, status, p, pi, lkf, cv}
 	err := i.Install()
 	require.NoError(t, err)
 	require.Equal(t, 2, statusReporters[0].(*execution.MockStatusReporter).RecipeInstalledCallCount)
@@ -729,7 +731,7 @@ func TestInstall_TargetedInstallInfraAgent_NoInfraAgentDuplicate(t *testing.T) {
 
 	v = validation.NewMockRecipeValidator()
 
-	i := RecipeInstaller{ic, d, l, mv, f, e, v, ff, status, p, pi, lkf}
+	i := RecipeInstaller{ic, d, l, mv, f, e, v, ff, status, p, pi, lkf, cv}
 	err := i.Install()
 	require.NoError(t, err)
 	require.Equal(t, 1, statusReporters[0].(*execution.MockStatusReporter).RecipeInstalledCallCount)
@@ -755,7 +757,7 @@ func TestInstall_TargetedInstall_SkipInfra(t *testing.T) {
 
 	v = validation.NewMockRecipeValidator()
 
-	i := RecipeInstaller{ic, d, l, mv, f, e, v, ff, status, p, pi, lkf}
+	i := RecipeInstaller{ic, d, l, mv, f, e, v, ff, status, p, pi, lkf, cv}
 	err := i.Install()
 	require.NoError(t, err)
 	require.Equal(t, 0, statusReporters[0].(*execution.MockStatusReporter).RecipeInstalledCallCount)
@@ -786,7 +788,7 @@ func TestInstall_TargetedInstall_SkipInfraDependency(t *testing.T) {
 
 	v = validation.NewMockRecipeValidator()
 
-	i := RecipeInstaller{ic, d, l, mv, f, e, v, ff, status, p, pi, lkf}
+	i := RecipeInstaller{ic, d, l, mv, f, e, v, ff, status, p, pi, lkf, cv}
 	err := i.Install()
 	require.NoError(t, err)
 	require.Equal(t, 1, statusReporters[0].(*execution.MockStatusReporter).RecipeInstalledCallCount)
@@ -833,7 +835,7 @@ func TestInstall_GuidReport(t *testing.T) {
 	// Test for NEW_RELIC_CLI_VERSION
 	os.Setenv("NEW_RELIC_CLI_VERSION", "testversion0.0.1")
 
-	i := RecipeInstaller{ic, d, l, mv, f, e, v, ff, status, p, pi, lkf}
+	i := RecipeInstaller{ic, d, l, mv, f, e, v, ff, status, p, pi, lkf, cv}
 	err := i.Install()
 	require.NoError(t, err)
 	require.Equal(t, 2, v.ValidateCallCount)
