+++ internal/install/scenario_builder.go
@@ -1,9 +1,7 @@
 package install
 
 import (
-	"github.com/sirupsen/logrus"
-	log "github.com/sirupsen/logrus"
-
+	"github.com/newrelic/newrelic-cli/internal/diagnose"
 	"github.com/newrelic/newrelic-cli/internal/install/discovery"
 	"github.com/newrelic/newrelic-cli/internal/install/execution"
 	"github.com/newrelic/newrelic-cli/internal/install/recipes"
@@ -16,22 +14,14 @@ import (
 type TestScenario string
 
 const (
-	Basic               TestScenario = "BASIC"
-	LogMatches          TestScenario = "LOG_MATCHES"
-	Fail                TestScenario = "FAIL"
-	StitchedPath        TestScenario = "STITCHED_PATH"
-	Canceled            TestScenario = "CANCELED"
-	DisplayExplorerLink TestScenario = "DISPLAY_EXPLORER_LINK"
+	Basic TestScenario = "BASIC"
+	Fail  TestScenario = "FAIL"
 )
 
 var (
 	TestScenarios = []TestScenario{
 		Basic,
-		LogMatches,
 		Fail,
-		StitchedPath,
-		Canceled,
-		DisplayExplorerLink,
 	}
 	emptyResults = []nrdb.NRDBResult{
 		map[string]interface{}{
@@ -70,16 +60,8 @@ func (b *ScenarioBuilder) BuildScenario(s TestScenario) *RecipeInstaller {
 	switch s {
 	case Basic:
 		return b.Basic()
-	case LogMatches:
-		return b.LogMatches()
 	case Fail:
 		return b.Fail()
-	case StitchedPath:
-		return b.StitchedPath()
-	case Canceled:
-		return b.CanceledInstall()
-	case DisplayExplorerLink:
-		return b.DisplayExplorerLink()
 	}
 
 	return nil
@@ -98,7 +80,10 @@ func (b *ScenarioBuilder) Basic() *RecipeInstaller {
 	c := validation.NewMockNRDBClient()
 	c.ReturnResultsAfterNAttempts(emptyResults, nonEmptyResults, 2)
 	v := validation.NewPollingRecipeValidator(c)
+	cv := diagnose.NewMockConfigValidator()
+	mv := discovery.NewEmptyManifestValidator()
 
+	lkf := NewMockLicenseKeyFetcher()
 	pf := discovery.NewRegexProcessFilterer(rf)
 	ff := recipes.NewRecipeFileFetcher()
 	d := discovery.NewPSUtilDiscoverer(pf)
@@ -117,6 +102,9 @@ func (b *ScenarioBuilder) Basic() *RecipeInstaller {
 		status:            statusRollup,
 		prompter:          p,
 		progressIndicator: s,
+		configValidator:   cv,
+		manifestValidator: mv,
+		licenseKeyFetcher: lkf,
 	}
 
 	i.InstallerContext = b.installerContext
@@ -137,7 +125,10 @@ func (b *ScenarioBuilder) Fail() *RecipeInstaller {
 	c := validation.NewMockNRDBClient()
 	c.ReturnResultsAfterNAttempts(emptyResults, nonEmptyResults, 2)
 	v := validation.NewPollingRecipeValidator(c)
+	cv := diagnose.NewMockConfigValidator()
+	mv := discovery.NewEmptyManifestValidator()
 
+	lkf := NewMockLicenseKeyFetcher()
 	pf := discovery.NewRegexProcessFilterer(rf)
 	ff := recipes.NewRecipeFileFetcher()
 	d := discovery.NewPSUtilDiscoverer(pf)
@@ -156,168 +147,9 @@ func (b *ScenarioBuilder) Fail() *RecipeInstaller {
 		status:            statusRollup,
 		prompter:          p,
 		progressIndicator: pi,
-	}
-
-	i.InstallerContext = b.installerContext
-
-	return &i
-}
-
-func (b *ScenarioBuilder) LogMatches() *RecipeInstaller {
-
-	// mock implementations
-	rf := setupRecipeFetcherGuidedInstall()
-	ers := []execution.StatusSubscriber{
-		execution.NewMockStatusReporter(),
-		execution.NewTerminalStatusReporter(),
-	}
-	slg := execution.NewConcreteSuccessLinkGenerator()
-	statusRollup := execution.NewInstallStatus(ers, slg)
-	c := validation.NewMockNRDBClient()
-	c.ReturnResultsAfterNAttempts(emptyResults, nonEmptyResults, 2)
-	v := validation.NewPollingRecipeValidator(c)
-	gff := discovery.NewMockFileFilterer()
-
-	pf := discovery.NewRegexProcessFilterer(rf)
-	ff := recipes.NewRecipeFileFetcher()
-	d := discovery.NewPSUtilDiscoverer(pf)
-	re := execution.NewGoTaskRecipeExecutor()
-	p := ux.NewPromptUIPrompter()
-	pi := ux.NewPlainProgress()
-
-	gff.FilterVal = []types.OpenInstallationLogMatch{
-		{
-			Name: "asdf",
-			File: "asdf",
-		},
-	}
-
-	i := RecipeInstaller{
-		discoverer:        d,
-		fileFilterer:      gff,
-		recipeFetcher:     rf,
-		recipeExecutor:    re,
-		recipeValidator:   v,
-		recipeFileFetcher: ff,
-		status:            statusRollup,
-		prompter:          p,
-		progressIndicator: pi,
-	}
-
-	i.InstallerContext = b.installerContext
-
-	return &i
-}
-
-func (b *ScenarioBuilder) StitchedPath() *RecipeInstaller {
-	// mock implementations
-	rf := setupRecipeFetcherStitchedPath()
-	ers := []execution.StatusSubscriber{
-		execution.NewMockStatusReporter(),
-		execution.NewTerminalStatusReporter(),
-	}
-	slg := execution.NewConcreteSuccessLinkGenerator()
-	statusRollup := execution.NewInstallStatus(ers, slg)
-	c := validation.NewMockNRDBClient()
-	c.ReturnResultsAfterNAttempts(emptyResults, nonEmptyResults, 2)
-	v := validation.NewPollingRecipeValidator(c)
-
-	pf := discovery.NewRegexProcessFilterer(rf)
-	ff := recipes.NewRecipeFileFetcher()
-	d := discovery.NewPSUtilDiscoverer(pf)
-	gff := discovery.NewGlobFileFilterer()
-	re := execution.NewGoTaskRecipeExecutor()
-	p := ux.NewPromptUIPrompter()
-	pi := ux.NewPlainProgress()
-
-	i := RecipeInstaller{
-		discoverer:        d,
-		fileFilterer:      gff,
-		recipeFetcher:     rf,
-		recipeExecutor:    re,
-		recipeValidator:   v,
-		recipeFileFetcher: ff,
-		status:            statusRollup,
-		prompter:          p,
-		progressIndicator: pi,
-	}
-
-	i.InstallerContext = b.installerContext
-
-	return &i
-}
-
-func (b *ScenarioBuilder) CanceledInstall() *RecipeInstaller {
-	// mock implementations
-	rf := setupRecipeCanceledInstall()
-	ers := []execution.StatusSubscriber{
-		execution.NewMockStatusReporter(),
-		execution.NewTerminalStatusReporter(),
-	}
-	slg := execution.NewConcreteSuccessLinkGenerator()
-	statusRollup := execution.NewInstallStatus(ers, slg)
-	c := validation.NewMockNRDBClient()
-	c.ReturnResultsAfterNAttempts(emptyResults, nonEmptyResults, 2)
-	v := validation.NewPollingRecipeValidator(c)
-
-	pf := discovery.NewRegexProcessFilterer(rf)
-	ff := recipes.NewRecipeFileFetcher()
-	d := discovery.NewPSUtilDiscoverer(pf)
-	gff := discovery.NewGlobFileFilterer()
-	re := execution.NewGoTaskRecipeExecutor()
-	p := ux.NewPromptUIPrompter()
-	pi := ux.NewPlainProgress()
-
-	i := RecipeInstaller{
-		discoverer:        d,
-		fileFilterer:      gff,
-		recipeFetcher:     rf,
-		recipeExecutor:    re,
-		recipeValidator:   v,
-		recipeFileFetcher: ff,
-		status:            statusRollup,
-		prompter:          p,
-		progressIndicator: pi,
-	}
-
-	i.InstallerContext = b.installerContext
-
-	return &i
-}
-
-func (b *ScenarioBuilder) DisplayExplorerLink() *RecipeInstaller {
-	log.StandardLogger().SetLevel(logrus.DebugLevel)
-
-	// mock implementations
-	rf := setupDisplayExplorerLink()
-	ers := []execution.StatusSubscriber{
-		execution.NewMockStatusReporter(),
-		execution.NewTerminalStatusReporter(),
-	}
-	slg := execution.NewConcreteSuccessLinkGenerator()
-	statusRollup := execution.NewInstallStatus(ers, slg)
-	c := validation.NewMockNRDBClient()
-	c.ReturnResultsAfterNAttempts(emptyResults, nonEmptyResults, 2)
-	v := validation.NewPollingRecipeValidator(c)
-
-	pf := discovery.NewRegexProcessFilterer(rf)
-	ff := recipes.NewRecipeFileFetcher()
-	d := discovery.NewPSUtilDiscoverer(pf)
-	gff := discovery.NewGlobFileFilterer()
-	re := execution.NewGoTaskRecipeExecutor()
-	p := ux.NewPromptUIPrompter()
-	pi := ux.NewPlainProgress()
-
-	i := RecipeInstaller{
-		discoverer:        d,
-		fileFilterer:      gff,
-		recipeFetcher:     rf,
-		recipeExecutor:    re,
-		recipeValidator:   v,
-		recipeFileFetcher: ff,
-		status:            statusRollup,
-		prompter:          p,
-		progressIndicator: pi,
+		configValidator:   cv,
+		manifestValidator: mv,
+		licenseKeyFetcher: lkf,
 	}
 
 	i.InstallerContext = b.installerContext
@@ -344,6 +176,11 @@ It is made up of a multi line string.
 				`,
 			},
 			ValidationNRQL: "test NRQL",
+			Install: `
+version: '3'
+tasks:
+  default:
+`,
 		},
 		{
 			Name:           "logs-integration",
@@ -355,70 +192,17 @@ It is made up of a multi line string.
 					File: "/var/lib/docker/containers/*/*.log",
 				},
 			},
+			Install: `
+version: '3'
+tasks:
+  default:
+`,
 		},
 	}
 	f.FetchRecommendationsVal = []types.OpenInstallationRecipe{
 		{
 			Name:           "recommended-recipe",
 			DisplayName:    "Recommended recipe",
-			ValidationNRQL: "test NRQL",
-		},
-	}
-
-	return f
-}
-
-func setupRecipeFetcherStitchedPath() recipes.RecipeFetcher {
-	f := recipes.NewMockRecipeFetcher()
-	f.FetchRecipeVals = []types.OpenInstallationRecipe{
-		{
-			Name:           "recommended-recipe",
-			DisplayName:    "Recommended recipe",
-			ValidationNRQL: "test NRQL",
-		},
-		{
-			Name:           "another-recommended-recipe",
-			DisplayName:    "Another Recommended recipe",
-			ValidationNRQL: "test NRQL",
-		},
-	}
-
-	return f
-}
-
-func setupRecipeCanceledInstall() recipes.RecipeFetcher {
-	f := recipes.NewMockRecipeFetcher()
-	f.FetchRecipeVals = []types.OpenInstallationRecipe{
-		{
-			Name:           "infrastructure-agent-installer",
-			DisplayName:    "Infrastructure Agent",
-			ValidationNRQL: "test NRQL",
-		},
-		{
-			Name:           "test-canceled-installation",
-			DisplayName:    "Test Canceled Installation",
-			ValidationNRQL: "test NRQL",
-		},
-	}
-
-	return f
-}
-
-func setupDisplayExplorerLink() recipes.RecipeFetcher {
-	f := recipes.NewMockRecipeFetcher()
-	f.FetchRecipeVals = []types.OpenInstallationRecipe{
-		{
-			Name:           "test-display-explorer-link",
-			DisplayName:    "Test Display Explorer Link",
-			ValidationNRQL: "test NRQL",
-			SuccessLinkConfig: types.OpenInstallationSuccessLinkConfig{
-				Type:   "explorer",
-				Filter: "\"`tags.language` = 'java'\"",
-			},
-		},
-		{
-			Name:           "infrastructure-agent-installer",
-			DisplayName:    "Infrastructure Agent",
 			ValidationNRQL: "test NRQL",
 		},
 	}
